[fix] corrigiendo errores por la relationship entre encuesta y variable
